refactor: comment mocking 파일 분리 (#42)
